make sure attributes are strings (#22)
This currently breaks inspec check when the default attribute isn't found and becomes nil.
```
Default postgresql configuration directory: does not exist. Postgresql may not be installed or we've misidentified the configuration directory.

Default postgresql data directory: does not exist. Postgresql may not be installed or we've misidentified the data directory.

/home/travis/.rvm/gems/ruby-2.3.3/bundler/gems/inspec-78c2a55945cb/lib/resources/postgres_conf.rb:29:in `dirname': can't convert Inspec::Attribute::DEFAULT_ATTRIBUTE to String (Inspec::Attribute::DEFAULT_ATTRIBUTE#to_str gives Inspec::Attribute::DEFAULT_ATTRIBUTE) (TypeError)
```